Left changes on ws10
