+++ nni/algorithms/compression/pytorch/pruning/__init__.py
@@ -5,11 +5,11 @@ from .finegrained_pruning_masker import *
 from .structured_pruning_masker import *
 from .one_shot_pruner import *
 from .iterative_pruner import *
-from .agp_pruner import *
+# from .agp_pruner import *
 from .lottery_ticket import LotteryTicketPruner
 from .simulated_annealing_pruner import SimulatedAnnealingPruner
 from .net_adapt_pruner import NetAdaptPruner
-from .admm_pruner import ADMMPruner
+# from .admm_pruner import ADMMPruner
 from .auto_compress_pruner import AutoCompressPruner
 from .sensitivity_pruner import SensitivityPruner
 from .amc import AMCPruner
