+++ nni/algorithms/compression/pytorch/pruning/iterative_pruner.py
@@ -5,15 +5,380 @@ import logging
 import torch
 from schema import And, Optional, SchemaError
 from nni.compression.pytorch.utils.config_validation import CompressorSchema
-from .one_shot_pruner import OneshotPruner, _StructuredFilterPruner
+from nni.compression.pytorch.compressor import Pruner
+from .constants import MASKER_DICT, MAX_EPOCHS
+from .structured_pruner import StructuredFilterPruner
 
-__all__ = ['SlimPruner', 'TaylorFOWeightFilterPruner', 'ActivationAPoZRankFilterPruner', 'ActivationMeanRankFilterPruner']
+__all__ = ['AGPPruner', 'ADMMPruner', 'SlimPruner', 'TaylorFOWeightFilterPruner', 'ActivationAPoZRankFilterPruner', 'ActivationMeanRankFilterPruner', ]
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 
-class SlimPruner(OneshotPruner):
+class IterativePruner(Pruner):
+    """
+    Prune model during the training process. 
+    """
+
+    def __init__(self, model, config_list, pruning_algorithm='level', optimizer=None, trainer=None, criterion=None, training_eopchs=1, **algo_kwargs):
+        """
+        Parameters
+        ----------
+        model : torch.nn.Module
+            Model to be pruned
+        config_list : list
+            List on pruning configs
+        pruning_algorithm: str
+            algorithms being used to prune model
+        optimizer: torch.optim.Optimizer
+            Optimizer used to train model
+        trainer: function
+            Function used to train the model.
+            Users should write this function as a normal function to train the Pytorch model
+            and include `model, optimizer, criterion, epoch, callback` as function arguments.
+        criterion: function
+            Function used to calculate the loss between the target and the output.
+        training_epochs : int
+            Totoal number of epochs for training.
+        algo_kwargs: dict
+            Additional parameters passed to pruning algorithm masker class
+        """
+        print("IterativePruner init Start")
+        super().__init__(model, config_list, optimizer)
+        self.set_wrappers_attribute("if_calculated", False)
+        self._trainer = trainer
+        self._criterion = criterion
+        self.masker = MASKER_DICT[pruning_algorithm](
+            model, self, **algo_kwargs)
+        print("IterativePruner init Done")
+
+    def compress(self):
+        training = self.bound_model.training
+        self.bound_model.train()
+        for epoch in range(MAX_EPOCHS):
+            self._trainer(self.bound_model, optimizer=self.optimizer,
+            criterion=self._criterion, epoch=epoch, callback=self._callback)
+            if epoch >= self.training_epochs:
+                break
+        self.bound_model.train(training)
+        self._get_threshold()
+
+        self.update_mask()
+        return self.bound_model
+
+
+    def _callback(self):
+        """
+        Callback function to do additonal optimization
+        """
+        pass
+
+
+class AGPPruner(IterativePruner):
+    """
+    Parameters
+    ----------
+    model : torch.nn.Module
+        Model to be pruned.
+    config_list : listlist
+        Supported keys:
+            - initial_sparsity: This is to specify the sparsity when compressor starts to compress.
+            - final_sparsity: This is to specify the sparsity when compressor finishes to compress.
+            - start_epoch: This is to specify the epoch number when compressor starts to compress, default start from epoch 0.
+            - end_epoch: This is to specify the epoch number when compressor finishes to compress.
+            - frequency: This is to specify every *frequency* number epochs compressor compress once, default frequency=1.
+    optimizer: torch.optim.Optimizer
+        Optimizer used to train model.
+    trainer: function
+        Function to train the model
+    pruning_algorithm: str
+        Algorithms being used to prune model,
+        choose from `['level', 'slim', 'l1', 'l2', 'fpgm', 'taylorfo', 'apoz', 'mean_activation']`, by default `level`
+    """
+
+    def __init__(self, model, config_list, optimizer, trainer, criterion, training_epochs=1, pruning_algorithm='level'):
+        super().__init__(model, config_list, optimizer=optimizer, trainer=trainer, criterion=criterion)
+        assert isinstance(optimizer, torch.optim.Optimizer), "AGP pruner is an iterative pruner, please pass optimizer of the model to it"
+        self.masker = MASKER_DICT[pruning_algorithm](model, self)
+        self.training_epochs = training_epochs
+        self.now_epoch = 0
+        self.set_wrappers_attribute("if_calculated", False)
+
+    def validate_config(self, model, config_list):
+        """
+        Parameters
+        ----------
+        model : torch.nn.Module
+            Model to be pruned
+        config_list : list
+            List on pruning configs
+        """
+        schema = CompressorSchema([{
+            'sparsity': And(float, lambda n: 0 <= n <= 1),
+            'frequency': And(int, lambda n: n > 0),
+            Optional('op_types'): [str],
+            Optional('op_names'): [str]
+        }], model, logger)
+
+        schema.validate(config_list)
+
+    def calc_mask(self, wrapper, wrapper_idx=None):
+        """
+        Calculate the mask of given layer.
+        Scale factors with the smallest absolute value in the BN layer are masked.
+        Parameters
+        ----------
+        wrapper : Module
+            the layer to instrument the compression operation
+        wrapper_idx: int
+            index of this wrapper in pruner's all wrappers
+        Returns
+        -------
+        dict | None
+            Dictionary for storing masks, keys of the dict:
+            'weight_mask':  weight mask tensor
+            'bias_mask': bias mask tensor (optional)
+        """
+
+        config = wrapper.config
+        freq = config.get('frequency', 1)
+
+        if wrapper.if_calculated:
+            return None
+
+        if not self.now_epoch % freq == 0:
+            return None
+
+        target_sparsity = self.compute_target_sparsity(config)
+        new_mask = self.masker.calc_mask(sparsity=target_sparsity, wrapper=wrapper, wrapper_idx=wrapper_idx)
+        if new_mask is not None:
+            wrapper.if_calculated = True
+
+        return new_mask
+
+    def compute_target_sparsity(self, config):
+        """
+        Calculate the sparsity for pruning
+        Parameters
+        ----------
+        config : dict
+            Layer's pruning config
+        Returns
+        -------
+        float
+            Target sparsity to be pruned
+        """
+
+        initial_sparsity = 0
+        end_epoch = self.training_epochs
+        freq = config.get('frequency', 1)
+        self.target_sparsity = final_sparsity = config.get('sparsity', 0)
+        
+        if initial_sparsity >= final_sparsity:
+            logger.warning('your initial_sparsity >= final_sparsity')
+            return final_sparsity
+
+        if end_epoch == 1 or end_epoch <= self.now_epoch:
+            return final_sparsity
+            
+        span = ((end_epoch - 1) // freq) * freq
+        assert span > 0
+        self.target_sparsity = (final_sparsity +
+                           (initial_sparsity - final_sparsity) *
+                           (1.0 - ((self.now_epoch) / span)) ** 3)
+        return self.target_sparsity
+
+    def update_epoch(self, epoch):
+        """
+        Update epoch
+        Parameters
+        ----------
+        epoch : int
+            current training epoch
+        """
+
+        if epoch > 0:
+            self.now_epoch = epoch
+            for wrapper in self.get_modules_wrapper():
+                wrapper.if_calculated = False
+
+    def compress(self):
+        for epoch in range(self.training_epochs):
+            self.update_epoch(epoch)
+            self._trainer(self.bound_model, optimizer=self.optimizer, criterion=self._criterion, epoch=epoch)
+            logger.info(f'sparsity is {self.target_sparsity:.2f} at epoch {epoch}')
+            self.update_mask()
+            self.get_pruned_weights()
+            
+        return self.bound_model
+        
+
+class ADMMPruner(IterativePruner):
+    """
+    A Pytorch implementation of ADMM Pruner algorithm.
+
+    Parameters
+    ----------
+    model : torch.nn.Module
+        Model to be pruned.
+    config_list : list
+        List on pruning configs.
+    trainer : function
+        Function used for the first subproblem.
+        Users should write this function as a normal function to train the Pytorch model
+        and include `model, optimizer, criterion, epoch, callback` as function arguments.
+        Here `callback` acts as an L2 regulizer as presented in the formula (7) of the original paper.
+        The logic of `callback` is implemented inside the Pruner,
+        users are just required to insert `callback()` between `loss.backward()` and `optimizer.step()`.
+        Example::
+
+            def trainer(model, criterion, optimizer, epoch, callback):
+                device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+                train_loader = ...
+                model.train()
+                for batch_idx, (data, target) in enumerate(train_loader):
+                    data, target = data.to(device), target.to(device)
+                    optimizer.zero_grad()
+                    output = model(data)
+                    loss = criterion(output, target)
+                    loss.backward()
+                    # callback should be inserted between loss.backward() and optimizer.step()
+                    if callback:
+                        callback()
+                    optimizer.step()
+    num_iterations : int
+        Total number of iterations.
+    training_epochs : int
+        Training epochs of the first subproblem.
+    row : float
+        Penalty parameters for ADMM training.
+    base_algo : str
+        Base pruning algorithm. `level`, `l1`, `l2` or `fpgm`, by default `l1`. Given the sparsity distribution among the ops,
+        the assigned `base_algo` is used to decide which filters/channels/weights to prune.
+
+    """
+
+    def __init__(self, model, config_list, optimizer, trainer, criterion, num_iterations=30, training_epochs=5, row=1e-4, base_algo='l1'):
+        self._base_algo = base_algo
+
+        super().__init__(model, config_list)
+
+        self._trainer = trainer
+        self._optimizer = optimizer
+        self._criterion = criterion
+        self._num_iterations = num_iterations
+        self._training_epochs = training_epochs
+        self._row = row
+
+        self.set_wrappers_attribute("if_calculated", False)
+        self.masker = MASKER_DICT[self._base_algo](self.bound_model, self)
+
+    def validate_config(self, model, config_list):
+        """
+        Parameters
+        ----------
+        model : torch.nn.Module
+            Model to be pruned
+        config_list : list
+            List on pruning configs
+        """
+
+        if self._base_algo == 'level':
+            schema = CompressorSchema([{
+                'sparsity': And(float, lambda n: 0 < n < 1),
+                Optional('op_types'): [str],
+                Optional('op_names'): [str],
+            }], model, logger)
+        elif self._base_algo in ['l1', 'l2', 'fpgm']:
+            schema = CompressorSchema([{
+                'sparsity': And(float, lambda n: 0 < n < 1),
+                'op_types': ['Conv2d'],
+                Optional('op_names'): [str]
+            }], model, logger)
+
+        schema.validate(config_list)
+
+    def _projection(self, weight, sparsity, wrapper):
+        '''
+        Return the Euclidean projection of the weight matrix according to the pruning mode.
+
+        Parameters
+        ----------
+        weight : tensor
+            original matrix
+        sparsity : float
+            the ratio of parameters which need to be set to zero
+        wrapper: PrunerModuleWrapper
+            layer wrapper of this layer
+
+        Returns
+        -------
+        tensor
+            the projected matrix
+        '''
+        wrapper_copy = copy.deepcopy(wrapper)
+        wrapper_copy.module.weight.data = weight
+        return weight.data.mul(self.masker.calc_mask(sparsity, wrapper_copy)['weight_mask'])
+
+    def compress(self):
+        """
+        Compress the model with ADMM.
+
+        Returns
+        -------
+        torch.nn.Module
+            model with specified modules compressed.
+        """
+        _logger.info('Starting ADMM Compression...')
+
+        # initiaze Z, U
+        # Z_i^0 = W_i^0
+        # U_i^0 = 0
+        Z = []
+        U = []
+        for wrapper in self.get_modules_wrapper():
+            z = wrapper.module.weight.data
+            Z.append(z)
+            U.append(torch.zeros_like(z))
+
+        optimizer = torch.optim.Adam(
+            self.bound_model.parameters(), lr=1e-3, weight_decay=5e-5)
+
+        # Loss = cross_entropy +  l2 regulization + \Sum_{i=1}^N \row_i ||W_i - Z_i^k + U_i^k||^2
+        # criterion = torch.nn.CrossEntropyLoss()
+
+        # callback function to do additonal optimization, refer to the deriatives of Formula (7)
+        def callback():
+            for i, wrapper in enumerate(self.get_modules_wrapper()):
+                wrapper.module.weight.data -= self._row * \
+                    (wrapper.module.weight.data - Z[i] + U[i])
+
+        # optimization iteration
+        for k in range(self._num_iterations):
+            logger.info('ADMM iteration : %d', k)
+
+            # step 1: optimize W with AdamOptimizer
+            for epoch in range(self._training_epochs):
+                self._trainer(self.bound_model, optimizer=optimizer,
+                              criterion=self._criterion, epoch=epoch, callback=callback)
+
+            # step 2: update Z, U
+            # Z_i^{k+1} = projection(W_i^{k+1} + U_i^k)
+            # U_i^{k+1} = U^k + W_i^{k+1} - Z_i^{k+1}
+            for i, wrapper in enumerate(self.get_modules_wrapper()):
+                z = wrapper.module.weight.data + U[i]
+                Z[i] = self._projection(z, wrapper.config['sparsity'], wrapper)
+                U[i] = U[i] + wrapper.module.weight.data - Z[i]
+
+        # apply prune
+        self.update_mask()
+
+        logger.info('Compression finished.')
+
+        return self.bound_model
+
+
+class SlimPruner(StructuredFilterPruner, IterativePruner):
     """
     Parameters
     ----------
@@ -38,11 +403,10 @@ class SlimPruner(OneshotPruner):
     """
 
     def __init__(self, model, config_list, optimizer, trainer, criterion, training_epochs=2, scale=0.0001):
-        super().__init__(model, config_list, pruning_algorithm='slim')
+        super().__init__(model, config_list=config_list, optimizer=optimizer, pruning_algorithm='slim')
         self.training_aware = True
         self.training_epochs = training_epochs
         self.scale = scale
-        self.optimizer = optimizer
         self._trainer = trainer
         self._criterion = criterion
 
@@ -72,7 +436,7 @@ class SlimPruner(OneshotPruner):
             wrapper.module.weight.grad.data.add_(self.scale * torch.sign(wrapper.module.weight.data))
 
 
-class TaylorFOWeightFilterPruner(_StructuredFilterPruner):
+class TaylorFOWeightFilterPruner(IterativePruner, StructuredFilterPruner):
     """
     Parameters
     ----------
@@ -109,7 +473,7 @@ class TaylorFOWeightFilterPruner(_StructuredFilterPruner):
         self.training_epochs = training_epochs
 
 
-class ActivationAPoZRankFilterPruner(_StructuredFilterPruner):
+class ActivationAPoZRankFilterPruner(IterativePruner, StructuredFilterPruner):
     """
     Parameters
     ----------
@@ -152,7 +516,7 @@ class ActivationAPoZRankFilterPruner(_StructuredFilterPruner):
         self.training_epochs = training_epochs
 
 
-class ActivationMeanRankFilterPruner(_StructuredFilterPruner):
+class ActivationMeanRankFilterPruner(IterativePruner, StructuredFilterPruner):
     """
     Parameters
     ----------
@@ -192,3 +556,9 @@ class ActivationMeanRankFilterPruner(_StructuredFilterPruner):
                          activation=activation, training_epochs=training_epochs)
         self.training_aware = True
         self.training_epochs = training_epochs
+
+
+
+
+
+
